Redefinition of some functionalities. No new ones added
