+++ include/player.h
@@ -13,18 +13,12 @@ typedef struct s_data
 }   t_player;
 
 t_player *init_player();
+
 void move_player(mlx_key_data_t keydata, void* param);
+
 void move_forward(t_player *player);
-void move_backward(t_player *player);
+void move_backwards(t_player *player);
 void move_left(t_player *player);
 void move_right(t_player *player);
-
-void draw_square(t_game *game);
-void ft_draw(t_game *game, int length);
-
-//Vision
-void draw_pov(mlx_image_t* img, t_player *player, int length);
-void draw_cone(mlx_image_t *img, t_player *player, int length);
-
 
 #endif
